+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - refactoring of the fixture function compilation (olegpidsadnyi)
 - related factory fix (olegpidsadnyi)
+- post_generation fixture dependency fixed (olegpidsadnyi)
 
 1.0.1
 -----